refactor(deps-restorer): split the frozen install into its subsystems
`install_frozen_lockfile.rs` carried two subsystems that share nothing
with the frozen install beyond being reachable from it:

- `build_phase` — running package build scripts once the tree exists
- `hoisted` — the flat-`node_modules` linker, its hoist plan, and the
  runtime-major probes that feed it

Both move out, leaving the frozen install itself — its inputs, error
type, and output — at 930 lines rather than 1830.

Re-exports are explicit and split by the visibility each item already
had, so the crate's public surface is unchanged and internal helpers
stay internal.

Pure code motion: no behaviour change, verified by comparing every
content line of the original against the union of the new modules.
